Use incremental rebalance protocol
KIP-429 adds support for incremental rebalancing of consumer groups.
What this means is that the rebalance events only carry the list of
partitions that were actually newly-assigned or revoked.  If a
currently-assigned partition does not change status during a rebalance,
then the rebalance events will no longer carry that partition.

This allows us to significantly simplify the rebalance callback:
1.  We no longer need to depend on revoke being called prior to every
    assign call.  This can be decoupled.
2.  The list of revoked partitions are truly the revoked ones; we don't
    need to worry about them being added immediately after.
3.  Similarly, we don't need to worry about assigned partitions already
    being active.  Any assigned partitions should always be new ones.

NOTE: Kafka doesn't support changing the type of a consumer group, so
existing consumer groups using non-incremental rebalancing will need to
be deleted and then recreated the new partition assignment strategy.