Fix unstable tests
Some tests are failing on github actions, there are not failing on local environment.
After some investigation is seems that in the code, so goroutines are freezed for some time.
Github actions have limited number of processors available, generally 1 or 2.
To try to simulate on local, we have to set the GOMAXPROCS=1.

1- fixed scheduler missed events code: some events where sent to a channel others were process directly by the scheduler resulting in desynchronization, in order to fix that send all events to the channel.
2- fix mini scheduler tests.